% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Gets the value of the WTW_API_KEY environment variable}
\usage{
get_api_key()
}
\value{
The value of the \code{WTW_API_KEY} environment variable.
}
\description{
Gets the value of the WTW_API_KEY environment variable
}
\examples{
\dontrun{get_api_key()}
}
