% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_languages.R
\name{get_available_languages}
\alias{get_available_languages}
\title{Gets available languages for use in the what3words API}
\usage{
get_available_languages()
}
\value{
A dataframe of available languages.
}
\description{
Gets available languages for use in the what3words API
}
\examples{
\dontshow{if (!identical(Sys.getenv("WTW_API_KEY"), "")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_available_languages()
\dontshow{\}) # examplesIf}
}
