% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_from_body.R
\name{get_error_from_body}
\alias{get_error_from_body}
\title{Returns the error message from the response}
\usage{
get_error_from_body(resp)
}
\arguments{
\item{resp}{An httr2 response.}
}
\value{
The error code and message as a single item character vector.
}
\description{
Returns the error message from the response
}
