% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_from_words.R
\name{coords_from_words}
\alias{coords_from_words}
\title{Returns the coordinates for a given what3words address}
\usage{
coords_from_words(words, full_details)
}
\arguments{
\item{full_details}{}
}
\description{
Returns the coordinates for a given what3words address
}
